Replaced combined module and import functions into separate objects
There were problems with the borrow checker if we put the module and import functions into the same vector, using the Function enum. This is because if we want to be able to allow side effects for the import functions (for example, setting a variable "function_was_called"), then we need to pass the vector as mutable. However, this became tricky while calling module functions, considering the vector would have to be borrow mutably twice: one time to obtain the function to be called, and a second time to pass it as an argument to the called function.

Instead, let's split them, because the module functions can actually be kept immutable always, whereas the import functions can be made mutable. When we call the module function, we can borrow the module functions twice, because they are immutable borrows, and when calling the import functions, we don't need to pass any arguments.