Now testing whether the import function is being called
We need to change the ImportFunction to have a boxed trait object (FnMut). Before, it was a normal function pointer. Unfortunately, those cannot hold state (which is needed for the function_was_called borrow), so we need to switch to a dynamic trait object. Additionally, since it is a mutable borrow, we need to use a FnMut trait.